Move xtgeoviz.mpl to xtgeoviz.plot
This will make the transition from xtgeo.plot easier by allowing users
to adapt to the same API, just in a separate package.
