+++ src/xtgeoviz/__init__.py
@@ -1,7 +1,7 @@
 # flake8: noqa
 from xtgeoviz.frontends import xsectplotting
 from xtgeoviz.frontends.xsectplotting import xsectplot
-from xtgeoviz.mpl.quickplot import quickplot
+from xtgeoviz.plot.quickplot import quickplot
 
 try:
     from .version import version
