--- src/xtgeoviz/mpl/__init__.py
@@ -1,6 +1,0 @@
-"""The XTGeoViz plot package, mpl (i.e. using matplotlib)"""
-from xtgeoviz.mpl.grid3d_slice import Grid3DSlice
-
-# flake8: noqa
-from xtgeoviz.mpl.xsection import XSection
-from xtgeoviz.mpl.xtmap import Map
