+++ src/xtgeoviz/plot/__init__.py
@@ -1,0 +1,5 @@
+"""The XTGeoViz plot package"""
+# flake8: noqa
+from .grid3d_slice import Grid3DSlice
+from .xsection import XSection
+from .xtmap import Map
