+++ src/xtgeoviz/plot/_colortables.py
