+++ src/xtgeoviz/plot/grid3d_slice.py
