+++ src/xtgeoviz/plot/quickplot.py
